topotato: rip out pytest fixture integration
It has broken due to upstream changes.  Again.  For the 5th? time.  This
time the fix is to rip all of this out and simplify.  This means losing
the ability to run non-topotato-iterator functions as tests, but keeping
that is plain not worth the effort of serious headache every time pytest
changes its internal APIs.

Signed-off-by: David Lamparter <equinox@diac24.net>