topotato: disable the /var shenanigans on FreeBSD
"over-mounting" /var/run on FreeBSD kills the ld.so cache, which then
makes libraries in /usr/local/lib not work, which then makes exabgp not
work.  We don't need the mount to begin with since the entire virtual
router's file system is writable on FreeBSD due to unionfs; just disable
this function for the time being.  If anything, it probably makes more
sense to start using overlayfs on Linux.

Signed-off-by: David Lamparter <equinox@diac24.net>